created a dashboard
