+++ app/TrackReferral/page.tsx
@@ -1,0 +1,1 @@
+<h1>Track Referral Status </h1>
